# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MonitorMonitorType = typeof MonitorMonitorType[keyof typeof MonitorMonitorType];


export const MonitorMonitorType = {
  Product_Price: 'Product Price',
  Product_Availability: 'Product Availability',
  Product_Catalog: 'Product Catalog',
  Website_Content: 'Website Content',
} as const;

export type MonitorStatus = typeof MonitorStatus[keyof typeof MonitorStatus];


export const MonitorStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export type MonitorCheckSource = typeof MonitorCheckSource[keyof typeof MonitorCheckSource];


export const MonitorCheckSource = {
  demo: 'demo',
  live: 'live',
} as const;

export interface Monitor {
  id: string;
  name: string;
  websiteUrl: string;
  monitorType: MonitorMonitorType;
  status: MonitorStatus;
  checkSource: MonitorCheckSource;
  /** @nullable */
  lastValue: string | null;
  createdAt: string;
  lastChecked: string;
}

export type ChangeSeverity = typeof ChangeSeverity[keyof typeof ChangeSeverity];


export const ChangeSeverity = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export type ChangeDataSource = typeof ChangeDataSource[keyof typeof ChangeDataSource];


export const ChangeDataSource = {
  demo: 'demo',
  live: 'live',
} as const;

export interface Change {
  id: string;
  monitorId: string;
  changeType: string;
  title: string;
  description: string;
  oldValue: string;
  newValue: string;
  severity: ChangeSeverity;
  sourceUrl: string;
  dataSource: ChangeDataSource;
  detectedAt: string;
}

export interface AIAnalysis {
  id: string;
  changeId: string;
  summary: string;
  whyItMatters: string;
  suggestedAction: string;
  createdAt: string;
}

export type EarlySignalWorkspaceMode = typeof EarlySignalWorkspaceMode[keyof typeof EarlySignalWorkspaceMode];


export const EarlySignalWorkspaceMode = {
  database: 'database',
  demo: 'demo',
} as const;

export interface EarlySignalWorkspace {
  mode: EarlySignalWorkspaceMode;
  monitors: Monitor[];
  changes: Change[];
  analyses: AIAnalysis[];
}

export type MonitorInputMonitorType = typeof MonitorInputMonitorType[keyof typeof MonitorInputMonitorType];


export const MonitorInputMonitorType = {
  Product_Price: 'Product Price',
} as const;

export interface MonitorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  websiteUrl: string;
  monitorType: MonitorInputMonitorType;
}

export type MonitorStatusInputStatus = typeof MonitorStatusInputStatus[keyof typeof MonitorStatusInputStatus];


export const MonitorStatusInputStatus = {
  active: 'active',
  paused: 'paused',
} as const;

export interface MonitorStatusInput {
  status: MonitorStatusInputStatus;
}

export interface MonitorCheckResult {
  success: boolean;
  message: string;
  changeDetected: boolean;
  /** @nullable */
  value: string | null;
  monitor: Monitor;
}

export interface ErrorResponse {
  error: string;
}