# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: earlysignal
    description: EarlySignal AI demo workspace operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /earlysignal/workspace:
    get:
      operationId: getEarlySignalWorkspace
      tags: [earlysignal]
      summary: Get the EarlySignal workspace
      description: Returns database-backed workspace data, or the centralized demo dataset when DEMO_MODE is enabled or the database is empty.
      responses:
        "200":
          description: EarlySignal workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarlySignalWorkspace"
  /earlysignal/monitors:
    post:
      operationId: createEarlySignalMonitor
      tags: [earlysignal]
      summary: Create a Product Price monitor
      description: Creates a monitor in the EarlySignal database. This MVP only accepts Product Price.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorInput"
      responses:
        "201":
          description: Created monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          description: Invalid monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /earlysignal/monitors/{id}/status:
    patch:
      operationId: updateEarlySignalMonitorStatus
      tags: [earlysignal]
      summary: Update monitor status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorStatusInput"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated monitor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "404":
          description: Monitor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /check/{id}:
    post:
      operationId: checkEarlySignalMonitor
      summary: Check a monitor now
      description: Fetches the monitor website and records a price change when the extracted value differs from its previous value.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Check completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorCheckResult"
        "404":
          description: Monitor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Monitor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        websiteUrl:
          type: string
        monitorType:
          type: string
          enum: [Product Price, Product Availability, Product Catalog, Website Content]
        status:
          type: string
          enum: [active, paused]
        checkSource:
          type: string
          enum: [demo, live]
        lastValue:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        lastChecked:
          type: string
          format: date-time
      required: [id, name, websiteUrl, monitorType, status, checkSource, lastValue, createdAt, lastChecked]
    Change:
      type: object
      properties:
        id:
          type: string
        monitorId:
          type: string
        changeType:
          type: string
        title:
          type: string
        description:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        severity:
          type: string
          enum: [Low, Medium, High]
        sourceUrl:
          type: string
        dataSource:
          type: string
          enum: [demo, live]
        detectedAt:
          type: string
          format: date-time
      required: [id, monitorId, changeType, title, description, oldValue, newValue, severity, sourceUrl, dataSource, detectedAt]
    AIAnalysis:
      type: object
      properties:
        id:
          type: string
        changeId:
          type: string
        summary:
          type: string
        whyItMatters:
          type: string
        suggestedAction:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, changeId, summary, whyItMatters, suggestedAction, createdAt]
    EarlySignalWorkspace:
      type: object
      properties:
        mode:
          type: string
          enum: [database, demo]
        monitors:
          type: array
          items:
            $ref: "#/components/schemas/Monitor"
        changes:
          type: array
          items:
            $ref: "#/components/schemas/Change"
        analyses:
          type: array
          items:
            $ref: "#/components/schemas/AIAnalysis"
      required: [mode, monitors, changes, analyses]
    MonitorInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        websiteUrl:
          type: string
          minLength: 1
        monitorType:
          type: string
          enum: [Product Price]
      required: [name, websiteUrl, monitorType]
    MonitorStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, paused]
      required: [status]
    MonitorCheckResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        changeDetected:
          type: boolean
        value:
          type: [string, "null"]
        monitor:
          $ref: "#/components/schemas/Monitor"
      required: [success, message, changeDetected, value, monitor]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]