# lib/api-zod/src/generated/types/change.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeDataSource } from './changeDataSource';
import type { ChangeSeverity } from './changeSeverity';

export interface Change {
  id: string;
  monitorId: string;
  changeType: string;
  title: string;
  description: string;
  oldValue: string;
  newValue: string;
  severity: ChangeSeverity;
  sourceUrl: string;
  dataSource: ChangeDataSource;
  detectedAt: Date;
}