# lib/api-zod/src/generated/types/monitor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MonitorCheckSource } from './monitorCheckSource';
import type { MonitorMonitorType } from './monitorMonitorType';
import type { MonitorStatus } from './monitorStatus';

export interface Monitor {
  id: string;
  name: string;
  websiteUrl: string;
  monitorType: MonitorMonitorType;
  status: MonitorStatus;
  checkSource: MonitorCheckSource;
  /** @nullable */
  lastValue: string | null;
  createdAt: Date;
  lastChecked: Date;
}