# lib/api-zod/src/generated/types/monitorCheckResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Monitor } from './monitor';

export interface MonitorCheckResult {
  success: boolean;
  message: string;
  changeDetected: boolean;
  /** @nullable */
  value: string | null;
  monitor: Monitor;
}